# server/swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/sessions":
    post:
      summary: Creates an authenticated session
      tags:
      - Session
      operationId: authenticate
      parameters: []
      responses:
        '200':
          description: session authenticated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/user_dto"
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/error_dto"
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/phone_confirmation"
    get:
      summary: get session
      tags:
      - Session
      operationId: getCurrentSession
      responses:
        '200':
          description: session info
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/user_dto"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/error_dto"
    delete:
      summary: Deletes an auth session
      tags:
      - Session
      operationId: logout
      responses:
        '200':
          description: session deleted
          content: {}
        '401':
          description: unauthorized
          content: {}
servers:
- url: https://{defaultHost}/api/v1
  variables:
    defaultHost:
      default: api.levimurray.dev
components:
  schemas:
    error_dto:
      type: object
      properties:
        message:
          type: string
    user_dto:
      type: object
      properties:
        id:
          type: string
        phone_number:
          type: string
        profile:
          "$ref": "#/components/schemas/profile"
        complete_profile:
          type: boolean
        complete_name:
          type: boolean
        complete_address:
          type: boolean
        complete_billing:
          type: boolean
      required:
      - id
      - phone_number
      - complete_profile
      - complete_name
      - complete_address
      - complete_billing,
      - profile
...
    phone_confirmation:
      type: object
      properties:
        phone_number:
          type: string
        confirmation_pin:
          type: string
      required:
      - phone_number
      - confirmation_pin
    profile:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        address:
          "$ref": "#/components/schemas/profile_address"
      required:
      - first_name
      - last_name
      - address
    profile_address:
      type: object
      properties:
        street_number:
          type: string
        street_name:
          type: string
        unit_number:
          type: string
        city:
          type: string
        province:
          type: string
        postal_code:
          type: string
      required:
      - street_number,
      - street_name,
      - unit_number,
      - city,
      - province,
      - postal_code
...